refactor(device-api): split flagcx_device.h into core and internal headers
